+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##  [0.0.4] - Apr 7th, 2018
+
+*   Dependencies, especially [shadowing](https://www.npmjs.com/package/shadowing), upgraded.
+*   Command option `--longurl <style>` added.
+
 ##  [0.0.3] - Mar 21st, 2018
 
 *   Command option `--verbose | -v` added.